Use new AssertNotImplemented() instead of Assert (false);//NYI Cleanup (and use more) qPlatform_Windows stuff in many places - replace older qPlatform_Win32/_WIN32 Use #if             __STDC_WANT_SECURE_LIB__ in several places to control use of_s functions (See http://www.open-std.org/jtc1/sc22/wg14/www/docs/n1079.htm)

+++ Library/Sources/Stroika/Foundation/Execution/Event.cpp
@@ -18,5 +18,5 @@ using	namespace	Stroika::Foundation::Execution;
  ********************************************************************************
  */
 #if		qTrack_ThreadUtils_HandleCounts
-LONG	Event::sCurAllocatedHandleCount		=	0;
+uint32_t	Event::sCurAllocatedHandleCount		=	0;
 #endif
